requirements.txt
streamlit
supabase
python-dotenv
selenium
beautifulsoup4
